Criação dos registros par importação ericksson

+++ POSapB1/Zopone.AddOn.PO/Model/SQL/Procedures/ZPN_SP_LOGIMPORTACAOPO.sql
@@ -1,6 +1,6 @@
 ﻿create procedure ZPN_SP_LOGIMPORTACAOPO
 (
-	@po_id int, 
+	@po_id numeric, 
 	@MensagemLog varchar(5000)
 )
 as
